Heap should use uint32_t all the way, no 'int' inside
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cryptominisat/branches/gaussian_differentmatrix@669 6c43cb19-1bce-4953-9960-ac6414807ffa

+++ Solver/RestartTypeChooser.cpp
@@ -70,7 +70,7 @@ void RestartTypeChooser::calcHeap()
     std::cout << "First vars:" << std::endl;
     #endif
     Heap<Solver::VarOrderLt> tmp(S->order_heap);
-    uint thisTopX = std::min(S->order_heap.size(), (int)topX);
+    uint thisTopX = std::min(S->order_heap.size(), topX);
     for (uint i = 0; i != thisTopX; i++) {
         #ifdef VERBOSE_DEBUG
         std::cout << tmp.removeMin()+1 << ", ";
